[DT][NFC] Internalize transposeNarrowN logic to LayoutAttrInterface Impl
Whether apply transposion from narrow-N to narrow-M is backend
implementation details, and we do not need to expose it to the type
converter. The encoding itself has enough information, like indexing
maps, narrow dimensions, etc., to infer the shapes and encoding info.
Instead of updating the RankedTensorTypes and the attached encoding in
type converter, we can just cook the logic in `getEncodingInfo` methods.
From the encoding, we know that whether it is narrow-N case, and we can
update the MaterializeEncodingInfo correspondingly. The type converter
can infer the transposed tensor type from it. Thus, we can simplify the
logic in the type conversion.

The documentation of `transposeNarrowN` is moved to
[CPU|GPU]EncodingExternalModels.cpp because all the implementaion
locates at the files.
